Make input field component more reusable
